HP-182 Enable importing `helusers.jwt` without Django REST Framework
The helusers.jwt is a nice package name. But in order to avoid an
ImportError when using it rest_framework and rest_framework_jwt packages
were required. Not anymore. Now it's possible to add new stuff into that
package.
